nuke some deprecated stuff
