+++ packages/react-query/src/ssg/index.ts
@@ -1,4 +1,2 @@
-export { createSSGHelpers } from './ssg';
-export type { CreateSSGHelpersOptions } from './ssg';
 export { createProxySSGHelpers } from './ssgProxy';
 export type { DecoratedProcedureSSGRecord } from './ssgProxy';
