+++ packages/server/src/core/types.ts
@@ -1,4 +1,3 @@
-import { inferObservableValue } from '../observable';
 import { inferTransformedProcedureOutput } from '../shared';
 import { AnyProcedure, ProcedureArgs } from './procedure';
 import { AnyRouter, AnyRouterDef, Router } from './router';
@@ -36,16 +35,6 @@ export type inferProcedureParams<TProcedure> = TProcedure extends AnyProcedure
   : never;
 export type inferProcedureOutput<TProcedure> =
   inferProcedureParams<TProcedure>['_output_out'];
-
-/**
- * @deprecated will be removed in next major as it's v9 stuff
- */
-export type inferSubscriptionOutput<
-  TRouter extends AnyRouter,
-  TPath extends keyof TRouter['_def']['subscriptions'] & string,
-> = inferObservableValue<
-  inferProcedureOutput<TRouter['_def']['subscriptions'][TPath]>
->;
 
 export type inferProcedureClientError<TProcedure extends AnyProcedure> =
   inferProcedureParams<TProcedure>['_config']['errorShape'];
