+++ packages/tests/server/react/useMutation.test.tsx
@@ -41,15 +41,6 @@ const ctx = konn()
             date: new Date(),
           })),
       }),
-      /**
-       * @deprecated
-       */
-      deprecatedRouter: t.router({
-        /**
-         * @deprecated
-         */
-        deprecatedProcedure: t.procedure.query(() => '..'),
-      }),
     });
 
     return getServerAndReactClient(appRouter);
